requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart
streamlit
requests
scikit-learn
pandas
matplotlib
seaborn